requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreschema==0.0.4
Django==4.2.2
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
factory-boy==3.2.1
Faker==18.11.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
